Use float in position value

+++ playerbot/strategy/values/PositionValue.h
@@ -8,11 +8,11 @@ namespace ai
     public:
         PositionEntry() : valueSet(false), x(0), y(0), z(0), mapId(0) {}
         PositionEntry(const PositionEntry &other) : valueSet(other.valueSet), x(other.x), y(other.y), z(other.z), mapId(other.mapId) {}
-        void Set(double x, double y, double z, uint32 mapId) { this->x = x; this->y = y; this->z = z; this->mapId = mapId; this->valueSet = true; }
+        void Set(float x, float y, float z, uint32 mapId) { this->x = x; this->y = y; this->z = z; this->mapId = mapId; this->valueSet = true; }
         void Reset() { valueSet = false; }
         bool isSet() { return valueSet; }
 
-        double x, y, z;
+        float x, y, z;
         bool valueSet;
         uint32 mapId;
     };
